+++ src/main/java/org/lastbamboo/common/stun/stack/message/attributes/StunAttributesFactoryImpl.java
@@ -58,12 +58,12 @@ public class StunAttributesFactoryImpl implements StunAttributesFactory
         
         final ByteBuffer bodyBuf = ByteBuffer.wrap(body);
         
-        if (!StunAttributeType.hasAttribute((short) type))
+        if (!StunAttributeType.hasAttribute(type))
             {
             LOG.debug("Did not recognize attribute type: "+type);
             return;
             }
-        final StunAttributeType enumType = StunAttributeType.convert((short) type);
+        final StunAttributeType enumType = StunAttributeType.convert(type);
         try
             {
             final StunAttribute attribute = createAttribute(enumType, bodyBuf);
